debugging netcdf error 13 issue

+++ soilgrids/combine_and_transform_raw_data.jl
@@ -60,7 +60,7 @@ function create_combined_data!(data, files, attrib, transform, outfilepath)
     # get parameter values at each layer
     read_nc_data!(data, files, filedir)
     # Replace missing with 0
-    data[typeof.(data) .== Missing] .= 0
+    data[typeof.(data) .== Missing] .= Float32(0)
     data .= transform(data)
     write_nc_out(data, lat, lon, z, attrib, outfilepath)
     Base.mv(outfilepath, joinpath(outputdir, outfilepath))
